catch exceptions thrown by the authenticate method
For example when a POST to the login form doesn't contain username and
password but the authentication store requires them like
Catalyst::Authentication::Store::DBIx::Class.

Without this the validate_form method of HTML::FormHandler stops and never
gets to the point where it clears the 'posted' from attribute leading to
an authentication attempt even if the login action is called for the first
time and no user can login to the app any more.